requirements.txt
# KDP Local Automation Dependencies

# Core
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# PDF Generation
fpdf2>=2.7.0

# LLM Backends
anthropic>=0.18.0  # Claude API

# Testing
pytest>=7.4.0
hypothesis>=6.92.0  # Property-based testing

# Optional: Development
black>=23.0.0
mypy>=1.0.0
